… Cargo.lock
… Cargo.toml
… DEVELOPMENT.md
… crates/gitbutler-branch-actions/tests/extra/mod.rs
… crates/gitbutler-branch-actions/tests/virtual_branches/convert_to_real_branch.rs
… crates/gitbutler-branch-actions/tests/virtual_branches/mod.rs
… crates/gitbutler-branch-actions/tests/virtual_branches/update_base_branch.rs
… crates/gitbutler-diff/src/lib.rs
… crates/gitbutler-edit-mode/src/lib.rs
… crates/gitbutler-oplog/Cargo.toml
… crates/gitbutler-oplog/src/oplog.rs
… crates/gitbutler-repo/src/lib.rs
… crates/gitbutler-repo/src/repository_ext.rs
… crates/gitbutler-tauri/Cargo.toml
… crates/gitbutler-tauri/src/logs.rs
… crates/gitbutler-tauri/src/main.rs
… crates/gitbutler-testsupport/src/suite.rs
… crates/gitbutler-testsupport/src/test_project.rs
… crates/gitbutler-watcher/Cargo.toml
… crates/gitbutler-watcher/src/handler.rs