An early PR to capture the research and first steps towards making `list_virtual_branches` faster by introducing `gix`.
Refactoring might happen here as well.

### Tasks

* [x] `list_virtual_branches` for CLI
* [x] figure out the best target for `list_virtual_branches` by profiling: **gitbutler_diff:workdir()**
* [x] Can `diff::workdir()` be implemented without writing objects (but also, without skipping files probably?) in `git2`?
    - <img width="916" alt="Screenshot 2024-08-28 at 08 10 15" src="https://github.com/user-attachments/assets/6eb181b6-fd89-4d39-b2b2-3e153b371173">
    - It kind-of works as it's a bit faster, doesn't write objects. Also there is no control over big files, but that's more of a diff-API issue.
* [x] make 'big files check' faster (**1.9x**)
* [x] remove duplicate `diff::workdir()` call when responding to filesystem events (reuse the data instead) - makes it **800ms** faster
* [x] fix failing tests
* [x] bring big-file support back to read-only `diff::workdir()`
* [x] fix E2E test - maybe it's a sign that `get_workspace_head()` is required?

### Notes for the Reviewer

* ~~`recalculate_everything()` is about twice as fast in this PR as it is on Master.~~ I think I compared debug with release or something. This PR is merely equally fast for responses to filesystem changes, but will write much less to the ODB in the process as it avoids `index.add_all(*)`.
* `diff::workdir()` is read-only, but a little slower as it now has to find big files to skip first (in order to be mergeable). Overall it's still an improvement, but wants `gix` (in follow-up)

### Follow-Ups

* `gix` for `diff::workdir()`
    - a way to get hunks from a gix diff - right now no such abstraction exists but maybe it should to make everything easier and more flexible?
* replace `get_worktree_head()`
    - A fast way to merge multiple trees into an index, one after another to make `get_workspace_head` faster
    - Write an index

### Shortcomings of current implementation

- Current implementation creates objects that it isn't interested in, as a side-effect of `index::add(*)`

### Performance Results

Everything was taken from the GitLab repository, with a `--release` build.

#### workspace big files

**Before**

**731ms** with `git2`.

```
DEBUG       ┝━ should_auto_snapshot [ 778ms | 0.94% / 31.94% ] check_if_last_snapshot_older_than: 300s
DEBUG       │  ┝━ worktree_files_larger_than_limit_as_git2_ignore_rule [ 731ms | 30.01% ] worktree_dir: "/Users/byron/dev-gb/gitlab"
DEBUG       │  ┕━ branch_lines_since_snapshot [ 24.0ms | 0.98% ] head_sha: 
```

**After**

**389ms** for traversing 78k files, a **1.88x** improvement. It's faster as it won't have to do an actual status, it just finds untracked files. Of course, ideally one day there is no need to exclude untracked large files as untracked files as virtual branches won't 'soak up' the worktree like they do now.

```
DEBUG       ┝━ should_auto_snapshot [ 549ms | 0.96% / 25.52% ] check_if_last_snapshot_older_than: 300s
DEBUG       │  ┝━ worktree_files_larger_than_limit_as_git2_ignore_rule [ 389ms | 18.09% ] worktree_dir: "/Users/byron/dev-gb/gitlab"
DEBUG       │  ┕━ branch_lines_since_snapshot [ 139ms | 6.47% ] head_sha: 
```

#### duplicate workdir call

**Before**

```
INFO     handle [ 2.15s | 0.00% / 100.00% ] event: ProjectFileChange(d4abe90e-4fed-4d86-bb80-048e3740f003, bar)
INFO     ┕━ recalculate_everything [ 2.15s | 1.07% / 100.00% ] paths: 1
DEBUG       ┝━ workdir [ 698ms | 32.48% ] commit_oid: 1b885e82d93fe522a7c8a0128f57c61a12d2a4de
DEBUG       ┝━ should_auto_snapshot [ 570ms | 0.89% / 26.52% ] check_if_last_snapshot_older_than: 300s
DEBUG       │  ┝━ worktree_files_larger_than_limit_as_git2_ignore_rule [ 403ms | 18.75% ] worktree_dir: "/Users/byron/dev-gb/gitlab"
DEBUG       │  ┕━ branch_lines_since_snapshot [ 148ms | 6.87% ] head_sha: 1928e73219a69ead3324c7ad3d0d5ffeb18625fb
INFO        ┕━ calculate_virtual_branches [ 858ms | 1.34% / 39.93% ]
DEBUG          ┕━ list_virtual_branches [ 829ms | 5.39% / 38.58% ]
DEBUG             ┕━ get_applied_status [ 714ms | 1.06% / 33.20% ]
DEBUG                ┕━ workdir [ 691ms | 32.14% ] commit_oid: 1b885e82d93fe522a7c8a0128f57c61a12d2a4de
```

**After**

```
INFO     handle [ 1.65s | 0.00% / 100.00% ] event: ProjectFileChange(d4abe90e-4fed-4d86-bb80-048e3740f003, bar)
INFO     ┕━ recalculate_everything [ 1.65s | 1.55% / 100.00% ] paths: 1
DEBUG       ┝━ workdir [ 750ms | 45.48% ] commit_oid: 1b885e82d93fe522a7c8a0128f57c61a12d2a4de
DEBUG       ┝━ should_auto_snapshot [ 580ms | 1.20% / 35.22% ] check_if_last_snapshot_older_than: 300s
DEBUG       │  ┝━ worktree_files_larger_than_limit_as_git2_ignore_rule [ 420ms | 25.48% ] worktree_dir: "/Users/byron/dev-gb/gitlab"
DEBUG       │  ┕━ branch_lines_since_snapshot [ 141ms | 8.54% ] head_sha: 1928e73219a69ead3324c7ad3d0d5ffeb18625fb
INFO        ┕━ calculate_virtual_branches [ 293ms | 1.66% / 17.75% ]
DEBUG          ┕━ list_virtual_branches_cached [ 265ms | 7.54% / 16.09% ]
DEBUG             ┕━ get_applied_status_cached [ 141ms | 8.55% ]
```

#### Final Result

On `master`, unfortunately, this takes about 1,9s as well, which is largely due to `workdir` now actually being faster there. The overhead of doing the `ignore_large_files_in_diffs()` call is significant, and it eats up all wins.

```
2024-08-28T20:38:35.471023Z  INFO handle: crates/gitbutler-watcher/src/handler.rs:57: close time.busy=1.87s time.idle=7.54µs event=ProjectFileChange(a4dfef20-d58e-47d8-984d-9f78eb012c2a, that)
```

However, this is the basis `gix` to come in and do a status much faster, which is when everything will be consistently faster once again.

```
INFO     handle [ 1.91s | 0.00% / 100.00% ] event: ProjectFileChange(d4abe90e-4fed-4d86-bb80-048e3740f003, bar)
INFO     ┕━ recalculate_everything [ 1.91s | 1.23% / 100.00% ] paths: 1
DEBUG       ┝━ workdir [ 1.07s | 35.45% / 55.88% ] commit_oid: 1b885e82d93fe522a7c8a0128f57c61a12d2a4de
DEBUG       │  ┕━ ignore_large_files_in_diffs [ 390ms | 20.43% ] limit_in_bytes: 50000000
DEBUG       ┝━ should_auto_snapshot [ 549ms | 0.98% / 28.78% ] check_if_last_snapshot_older_than: 300s
DEBUG       │  ┝━ ignore_large_files_in_diffs [ 396ms | 20.75% ] limit_in_bytes: 33554432
DEBUG       │  ┕━ branch_lines_since_snapshot [ 134ms | 7.05% ] head_sha: 1928e73219a69ead3324c7ad3d0d5ffeb18625fb
INFO        ┕━ calculate_virtual_branches [ 269ms | 1.25% / 14.11% ]
DEBUG          ┕━ list_virtual_branches_cached [ 245ms | 6.02% / 12.86% ]
DEBUG             ┕━ get_applied_status_cached [ 131ms | 6.84% ]
```

It's notable though that technically, `diff::workdir` is now **25% slower** at least, and all that so it won't add everything to the worktree.
And I am quite aware that there are many locations that call `create_wd_tree()` which does exactly that, but I suppose it has to start somewhere.

### Performance 

* `list_virual_branches()` on the GitLab repo takes 1.16s, of which 802ms are spent on the workdir diff.
* `worktree::add(*)` is most time-consuming, but it's the dirwalk that really takes time. <img width="556" alt="Screenshot 2024-08-27 at 15 44 07" src="https://github.com/user-attachments/assets/858a7789-ca88-4346-a79c-d69612b50f66">